♻️ refactor(menu): category accordion list in its own component
